Clean up structure of app package
