+++ packages/app/src/App.tsx
@@ -8,7 +8,7 @@ import { ReflexContainer, ReflexElement, ReflexSplitter } from 'react-reflex';
 import styles from './App.module.css';
 import BreadcrumbsBar from './breadcrumbs/BreadcrumbsBar';
 import { type FeedbackContext } from './breadcrumbs/models';
-import { DimMappingProvider } from './dimension-mapper/store';
+import { DimMappingProvider } from './dim-mapping-store';
 import EntityLoader from './EntityLoader';
 import ErrorFallback from './ErrorFallback';
 import MetadataViewer from './metadata-viewer/MetadataViewer';
