--- packages/app/src/dimension-mapper/hooks.ts
@@ -1,20 +1,0 @@
-import { type DimensionMapping, getSliceSelection } from '@h5web/lib';
-import {
-  type ArrayShape,
-  type Dataset,
-  type ScalarShape,
-} from '@h5web/shared/hdf5-models';
-
-import { useDataContext } from '../providers/DataProvider';
-
-export function useValuesInCache(
-  ...datasets: (Dataset<ScalarShape | ArrayShape> | undefined)[]
-): (dimMapping: DimensionMapping) => boolean {
-  const { valuesStore } = useDataContext();
-  return (dimMapping) => {
-    const selection = getSliceSelection(dimMapping);
-    return datasets.every(
-      (dataset) => !dataset || valuesStore.has({ dataset, selection }),
-    );
-  };
-}
