+++ packages/app/src/vis-packs/core/matrix/MatrixVisContainer.tsx
@@ -5,8 +5,8 @@ import {
   assertPrintableType,
 } from '@h5web/shared/guards';
 
-import { useValuesInCache } from '../../../dimension-mapper/hooks';
-import { useDimMappingState } from '../../../dimension-mapper/store';
+import { useDimMappingState } from '../../../dim-mapping-store';
+import { useValuesInCache } from '../../../hooks';
 import visualizerStyles from '../../../visualizer/Visualizer.module.css';
 import { type VisContainerProps } from '../../models';
 import VisBoundary from '../../VisBoundary';
